requirements.txt
Django>=4.2,<5.0
djangorestframework
django-filter
psycopg2-binary
django-environ